fix: seperate description page
