+++ app/lib/description/account_info_page.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+import 'package:walk_guide/services/auth_service.dart';
+
+class AccountInfoPage extends StatelessWidget {
+  const AccountInfoPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('계정 정보')),
+      body: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 20),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: const [
+            CircleAvatar(
+              radius: 48,
+              backgroundImage: AssetImage('assets/images/profile.jpg'),
+            ),
+            SizedBox(height: 16),
+            Text('syprisilla', style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold)),
+            SizedBox(height: 32),
+            Row(
+              children: [
+                Icon(Icons.calendar_today_outlined, size: 20),
+                SizedBox(width: 8),
+                Text('가입한 날짜: 2017년 2월'),
+              ],
+            ),
+            SizedBox(height: 16),
+            Row(
+              children: [
+                Icon(Icons.email_outlined, size: 20),
+                SizedBox(width: 8),
+                Text('이메일: syprisilla@example.com'),
+              ],
+            ),
+          ],
+        ),
+      ),
+      bottomNavigationBar: Padding(
+        padding: const EdgeInsets.fromLTRB(20, 0, 20, 30),
+        child: SizedBox(
+          width: double.infinity,
+          height: 50,
+          child: ElevatedButton(
+            onPressed: () => signOut(context),
+            style: ElevatedButton.styleFrom(
+              backgroundColor: Colors.white,
+              foregroundColor: Colors.redAccent,
+              shape: RoundedRectangleBorder(
+                borderRadius: BorderRadius.circular(10),
+              ),
+            ),
+            child: const Text('로그아웃', style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold)),
+          ),
+        ),
+      ),
+    );
+  }
+}
